Optimize iso2022 map struct on x86-64

+++ src/iso2022.c
@@ -50,7 +50,7 @@
 #define INVALID_CODEPOINT 0xFFFD
 
 struct _vte_iso2022_map {
-	gulong from;
+	guint from;
 	gunichar to;
 };
